requirements.txt
absl-py==1.2.0
adamp==0.3.0
batchgenerators==0.24
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
cycler==0.11.0
Cython==3.0.8
dicom2nifti==2.4.3
easydict==1.10
efficientnet-pytorch==0.7.1
einops==0.7.0
filelock==3.13.1
fonttools==4.37.1
fsspec==2024.2.0
future==0.18.2
h5py==3.8.0
huggingface-hub==0.20.3
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.21.2
joblib==1.2.0
kiwisolver==1.4.4
linecache2==1.0.0
matplotlib==3.5.3
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
monai==1.3.0
munch==4.0.0
networkx==2.8.6
nibabel==4.0.2
nnunet==1.7.0
numpy==1.23.0
opencv-python==4.8.1.78
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pip==22.1.2
pretrainedmodels==0.7.4
protobuf==3.20.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydicom==2.3.0
pymeshlab==2022.2.post3
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-gdcm==3.0.19
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
safetensors==0.4.2
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
segmentation-models-pytorch==0.3.3
setuptools==63.4.1
SimpleITK==2.1.1.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
Surface-Distance-Based-Measures @ file:///public/bme/home/v-tanmh/Tooth_demo/surface-distance
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.9.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
traceback2==1.4.0
trimesh==3.17.1
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
unittest2==1.1.0
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
vtk==9.3.0
wheel==0.37.1
